print_adress

+++ format_and_print.c
@@ -18,7 +18,7 @@ int	format_and_print(va_list arg_list, char format)
 	if (format == 'X')
 		len += print_hex(va_arg(arg_list, int), 1);
 	if (format == 'p')
-		print_adress(va_arg(arg_list, int));
+		print_adress(va_arg(arg_list, unsigned long));
 	if (format == 'u')
 		return (0);
 	return (len);
